Develop

+++ builder/img/heroku-24-build.sh
@@ -63,5 +63,9 @@ apt-get install -y --no-install-recommends "${packages[@]}"
 
 rm -rf /root/*
 rm -rf /tmp/*
-rm -rf /var/cache/apt/archives/*.deb
-rm -rf /var/lib/apt/lists/*
+if ! mountpoint -q /var/cache/apt/archives 2>/dev/null; then
+  rm -rf /var/cache/apt/archives/* "/var/cache/apt/archives/partial"/*
+fi
+if ! mountpoint -q /var/lib/apt/lists 2>/dev/null; then
+  rm -rf /var/lib/apt/lists/*
+fi
